feat: allow to control the report's output directory
A new CLI flag that allows to save the report to a user-defined
directory instead of the default `target/llvm-cov-pretty` location.

Caution should be taken as there is no additional logic for non-empty
directories. The target directory will be fully wiped in the process of
generating the report, without prior confirmation.